+++ src/github.ts
@@ -1,5 +1,6 @@
 import * as github from '@actions/github'
-import * as pluginRetry from '@octokit/plugin-retry'
-import { Octokit } from './types.js'
+import { retry } from '@octokit/plugin-retry'
 
-export const getOctokit = (token: string): Octokit => github.getOctokit(token, undefined, pluginRetry.retry)
+export type Octokit = ReturnType<typeof github.getOctokit>
+
+export const getOctokit = (token: string): Octokit => github.getOctokit(token, {}, retry)
