+++ src/types.ts
@@ -1,7 +1,3 @@
-import * as github from '@actions/github'
-
-export type Octokit = ReturnType<typeof github.getOctokit>
-
 export type Issue = {
   owner: string
   repo: string
